+++ rdwt/admin/partials/display-overview.php
@@ -6,7 +6,7 @@ if ( ! function_exists( 'add_action' ) ) die() ;
 
 <div class="wrap rdwt-admin-wrap">
 
-  <h1 class="rdwt-title"><?php esc_html_e( 'Rdev WP Tools', 'rdwt' ); ?></h1>
+  <h1 class="rdwt-title"><?php esc_html_e( 'Rdev WP Tools', RDWT_DOMAIN ); ?></h1>
   <?php settings_errors(); ?>
   <h2 class="nav-tab-wrapper"></h2>
 
@@ -18,14 +18,18 @@ if ( ! function_exists( 'add_action' ) ) die() ;
       <div class="meta-box-sortables ui-sortable">
         <div id="rdwt-overview" class="postbox">
 
-          <h2><?php esc_html_e( 'Overview', 'rdwt' ); ?></h2>
+          <h2><?php esc_html_e( 'Overview', RDWT_DOMAIN ); ?></h2>
 
           <div class="panel overview">
             <p>
               <?php 
-              esc_html_e( 'Rdev WP Tools', 'rdwt' );
-              esc_html_e( ' is a collection of tools in a single bloat-less plugin.', 'rdwt' );
-              ?></p>
+              esc_html_e( 'Rdev WP Tools', RDWT_DOMAIN );
+              esc_html_e( ' is a collection of tools in a single bloat-less plugin.', RDWT_DOMAIN );
+              ?>
+            </p>
+            <ul>
+              <li><?php esc_html_e( 'Google Analytics tracking code', RDWT_DOMAIN ); ?></li>
+            </ul>
           </div>
 
         </div>