+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## v1.0.0
 
 * Initial Release.
+* Google Analytics 4 tracking code